Rename: DefaultTemplate nav 파일명 변경
